int32 64-bit kernel indexing fix + HVAB/HART2 configs + accumulated solver WIP
- Fix int32 overflow in D3Q27 CUDA kernels (q*N+idx / q*NxNyNz) that crashed at >79.5M cells/level (DGX fine L4=120M). 10 kernels -> 64-bit population indexing. Validated: 87M single-level repro pass, fine_mini 5-level bit-identical (T_lu unchanged), DGX fine OK.
- multi_level_grid: single-level f_prev guard.
- HVAB: NASA OVERFLOW C81 deck support (per-config polar_source), fine_watanabe baseline config, debug presets (fine_mini/ovf87); new config sets (alm/caradonna_tung/clf5605/cylinders/hart2/sgs_test/hvab).
- New d3q27/d2q9 kernels (dyn_smag/wale/macro/interpolated_wall), src/turbulence, utilities (tail_select/hover_fm_post/compare_spanwise).
- Accumulated multi-session solver WIP; config cleanup/removals included.

+++ src/kernels/interpolation_d3q27.py
@@ -34,7 +34,7 @@ void cubic_interp_3d(
     if (tid >= total) return;
 
     int NyNz = Ny * Nz;
-    int NxNyNz = Nx * NyNz;
+    long long NxNyNz = (long long)Nx * NyNz;   // 64-bit: q*NxNyNz overflows int32 at N>~79.5M
 
     int q  = tid / NxNyNz;
     int rem = tid - q * NxNyNz;
@@ -103,7 +103,7 @@ void cubic_interp_1d_x(
     int iy = rem / Nz;
     int iz = rem - iy * Nz;
 
-    int NxNyNz = Nx * NyNz;
+    long long NxNyNz = (long long)Nx * NyNz;   // 64-bit: q*NxNyNz overflows int32 at N>~79.5M
     #define FX(x) f[q * NxNyNz + (x) * NyNz + iy * Nz + iz]
 
     // Interior odd indices: 3, 5, ..., Nx-4
@@ -140,7 +140,7 @@ void cubic_interp_1d_y(
     if (tid >= total) return;
 
     int NyNz = Ny * Nz;
-    int NxNyNz = Nx * NyNz;
+    long long NxNyNz = (long long)Nx * NyNz;   // 64-bit: q*NxNyNz overflows int32 at N>~79.5M
 
     int q = tid / (Nx * Nz);
     int rem = tid - q * Nx * Nz;
@@ -180,7 +180,7 @@ void cubic_interp_1d_z(
     if (tid >= total) return;
 
     int NyNz = Ny * Nz;
-    int NxNyNz = Nx * NyNz;
+    long long NxNyNz = (long long)Nx * NyNz;   // 64-bit: q*NxNyNz overflows int32 at N>~79.5M
 
     int q = tid / (Nx * Ny);
     int rem = tid - q * Nx * Ny;
